adding asycblockingqueue

+++ management/src/main/groovy/galecino/VehicleService.groovy
@@ -242,7 +242,7 @@ abstract class VehicleService {
 
     }
 
-    private float map_range(float x, float X_min, float X_max, float Y_min, float Y_max) {
+    private float map_range(float x, int X_min, int X_max, int Y_min, int Y_max) {
     //    '''
     //Linear mapping between two ranges of values
     //'''
